+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## et-orbi 1.1.2  not yet released
 
+- Let EtOrbi.get_tzone understand "CST+0800"
 - Introduce EtOrbi.to_windows_tz (Asia/Kolkata to IST-5:30)